Move plot related code from Arena object to PlotsManager
